Random interactions and color phrase
Begin the random interactions based in chars, this allow a lot non-numeric characters interactions in the same map, which it's better to be bigger (since the 0-9 actual limitation by "world).

I am passing ThreadRng as dependency in map core too, this will be the strategy to propagate the random between the modules for now

The base modules are still messy, but in one moment they will follow the rust module structure

Only random phrases for now (but they are with color :D). The next steps:

- More than just random phrases, more interactions
- Keep the state of this interactions in the memory and sync this with the map