First interactions with monsters
Adding the monster in random interactions. There will be a list of monsters (0-N) in the same tile. The monsters are still very static and with fixed attributes.

But in this commit I am preparing it's coloured description and it's creation. The real interaction between player and monsters are coming (fight).

But first I need to provide more interactions input and draw the fighting system. Maybe in the next commit I can see some pillage.