First interactions between player and monsters
In theses commit, some memory references are made to keep the states mutables. This is the first of "arena loop", where the combats will happen.

I remove the vec of monsters, this is too much in this stage. Renaming `Interaction` to Map `Event` makes more sense, since I have a module called `interactions`.

I will put a "cache" like in random generated events using the three points of location: map index + x + y. This is the pointer to deaths of monster will be permanent.

BUT I still need to persist this truly for real in this hashmap. Probably the next step.

Renaming interaction to event

First interaction with human input between player and monsters

First monster defeat - need to persist in memory