Dead monster reference + monster types
This commit handle monsters death - the fight will not happen again in theses cases. For now, fights will not have "escape" behavior, they will be until the end.

In next step, I am thinking in some kind of loot interactive, but first I need to make the fight real: until now, only player have actions.

This PR adds too `MonsterType` trait, that will be the common interface to monsters.

The first one is the skeleton - the name fits with this trait for now :P

I will keeping the fights with more options and actions in the next steps.