Pre arena and real fights (with player death)
This commit adds "pre arena" function that will check if the player wants to start a fight. If the battle is started, it's until the dead, but now the player can ignore the enemy (this adds the possibility of aggro monster for free soon).

The arena also happens now in turns, based in higher agility. In this first version I am using only agility/strength/vital points in very very rudimentar way to validate the flow.

The base monster concept is also shaped to improvements in battle. This is the next step, keep the battles more "live" (even only with one kind of attack - `bash`)

The first class supported will be the fighter with obvious reasons... I will focus first in physical battles, monster types and improve the display of stats for a better user experience (classic prompt with the current stats, etc)