Arena module and some minor interactions
This commit separate `Arena` in a module, having ThreadRng (to random rolls in fights) and debug flag too.

A simple prompt of player stats is show during the fight too. I am still thinking what is the best way to inform this later.

The `render::render_mini_stats` is prepared to render monster virtual points too, but in terms of gameplay this can be obtained with some skills or other things