Supporting other monsters
This commit removes hard coded attributes in a more dynamic way, using the level as index in attributes creation. `Monster` and `MonsterType` probably will become just one thing soon, but for now the trait are working well for this fit.

Now, the game comes with a random approach and roguelike way: if the player chooses the wrong decision, death will come. I will add more random in this formula, so the results will be more and more unpredictable.