MonsterFactory to random encounters
This commit add a factory to generate monsters in events. Instead of adding them in vectors and randomize in runtime, this factory allows the generation in `generate_world` moment.

I remove the multiple insteractions behavior to a simple one (to do this change), because of this `Event` will receive only one monster at each moment. I will support multiple events with a more dynamic way again in the next commit.

Original monster levels are getting redudant in the approach of monster factory by level, but I will keep it for now. Maybe I remove it later.

Some other little adjusts in the player char (like color) and player position (to avoid "instant battles")  are made in this commit too