Encounter system draft
This commit adds the entire "encounter" factory in map and with a similar way of monsters. Encounters (for now) are everything not related to monsters and battles, but itens and any kind of stuff the player can interactive with - the first one is a potion.

Encounters can have "is_evil" effect - with a 50% of chance to not worked well as expected. For sure, this will be improved in the random seed too. The design is very similar with "monster factory" and works in a similar way.

In each dungeon floor, the player will find one encounter and one monster. I am thinking about X random encounters in some moment, or even fixed encounters, but I will make other improvements before this. Probably a mid solution is a fixed encounters using some fixed letters to keep things simples while I work in the game system. With encounter system, the game itself has the minimum to be playable.

In the next steps, maybe I will do some refactoring in arena module to separate these concepts - or just renaming modules to a better concept. And sure, looking to add more monsters and encounters.