Itens module
This commit replace the previous "encounter" module to a less generic and direct module. This will handle itens discovery, that can be weapons, consumable itens, fixed itens (cannot be picked), traps, etc.

The next improvements will be the weapons - they are the base of fights anyway. The current "potion" has a fixed item behavior, and I am changing it name before. Pickable/Consumable/Dropable itens gives a complexibility that I will not do now.

This is a commit to change this design to weapon first