Adding the draft of weapon type trait
This commit adds the draft of weapon type and some basic weapons and damage type: swords with slash, maces that bash and lances with piercing. This is very basic approach and not doing nothing for now. I will experiment different approaches to this trait WeaponType can be used in Item struct in a some generic way, but still without success in the way I want.

Renaming Encounter* to Item* to in `itens` module.