WeaponType trait boxed as generic type in Item struct
This commit adds `dyn WeaponType` in Item property weapon. I am using `Box` to store this values in heap, with static lifetime and manual clone trait implementation, since they will be added dynamically in levels.

This is exactly what I am looking for, but still need improvements. Some libs like `dyn-clone` does this under the hoods, but it's very interesting (and a good challenge) does this manually (ex: I get two stack overflow faults in different places when I was doing the implementation and now I know how to avoid them at all). I still need to improve `Debug` trait implementation - it is not working, but at least it's not throwing stack overflow in the stable version.

I will shape it better in the next steps. With this base working, I can look for add more features. This will works well to every generic behavior that I was looking for.