Option to Box<dyn WeaponType>
This commit adds the optional value to weapon types. In actual `Item` struct (with two kinds of itens: potions and weapons) I need to allow `None` in itens that is not weapons. I will add more weapons and provide a simple way to create new weapons - they are scattered in few structs with overlap of responsability.

Also, the player now can `pick` weapons, using `clone` method. I think it's nice for now to not deal of refs - just cloning is fair.

In the next steps I will keep improving the `itens` module until it is ok to add more itens and games rules.