Auto mini-stats, first command "stats" and improving WeaponType with trait Display
This commit has more focus in game play, showing the mini stats in each moviment (soon I will add some color in this display) and the first explict command "stats" that render the current attributes and equipment of player. This will be the base of more commands and input different than directions.

I also remove `Option` from `Player` weapon - it is required now (no need to match `Option` everywhere now). To this, I assume a unarmed player is using the weapon `BareHands`. Simple as this.

Adding a description in the `WeaponType` is nice to gain the advantages of native trait `Display` for free. This commit adds this implementation too. I will keep description in encounters too, to give more flexibiliy to describe itens and encounters.

I still need to do some adjusts in random minimap positions. But the whole dynamic `weapon` concept that I shape  will help a lot to other entities of players (armor, itens, etc).