Recursive generate_position function
Remove the previous implementation (using `while`) to a recursive function. This will allows multiple monsters/encounters position in the same map - and solves the "TODO" of last commit. I am also removing `description` of room... The first idea was of rooms with great description, but with all the random behavior, this don't make sense anymore - maybe it return in the future.