Creating Position2d struct to store x and y
Fixing other "TODO": the itens/monsters are using a HashMap based in string keys, like "X" and "?". This worked in the beginning of project to validate the encounter system, but with the evolution of map generation, etc this is not working well. In the random seed I want to put a range of a monster and itens by level, and to alllow this, I need to use as HashMap key a struct like Position2d.

With this, I just replaced all x/y logic of tuples of (usize, usize) to a struct. With the random position ok, the position 2d concept and other fixes in the last days, the multi monsters/itens in the same map generation is next.