Multiple monsters and encounters in map places
This is the feature that will changes everything. With this, our premise of "just one monster/one item by room" is not true anymore. This will allow a lot of more and more random games, with unpredictable maps by level. The main goal of previous features/refactorings (position2d, recursion in map placement, etc) is allow this dynamic/random/multiple encounters.

With this base I will improve the itens/monsters in the next steps - that limited position implementation was not nice.

Now, the game will start to get shape